Master Branch update from official repo

+++ src/rootweb.cpp
@@ -1,7 +1,5 @@
 #include <rootweb.hh>
 
-#include <messageLogger.h>
-
 int RootWImage::imageCounter_ = 0;
 std::map <std::string, int> RootWImage::imageNameCounter_;
 
@@ -78,7 +76,7 @@ ostream& RootWTable::dump(ostream& output) {
       }
       output << ">";
       output << myTableContent[make_pair(iRow, iCol)];
-      output << "</" << myColorCode << ">" << " ";
+      output << "</" << myCellCode << ">" << " ";
     }
     output << "</tr>";
   }
@@ -398,8 +396,13 @@ void RootWImage::saveSummaryLoop(TPad* basePad, std::string baseName, TFile* myT
     } else if (
 	       (myClass=="TProfile") ||
 	       (myClass=="TGraph") ||
+	       (myClass=="TGraphErrors") ||
+	       (myClass=="TH1I") ||
+	       (myClass=="TH1F") ||
 	       (myClass=="TH1D") ||
 	       (myClass=="TH2C") ||
+	       (myClass=="TH2I") ||
+	       (myClass=="TH2F") ||
 	       (myClass=="TH2D") ||
 	       (myClass=="THStack") 
 	       ) {
@@ -417,10 +420,12 @@ void RootWImage::saveSummaryLoop(TPad* basePad, std::string baseName, TFile* myT
 	       (myClass=="TLine") ||
 	       (myClass=="TPaveText") ||
 	       (myClass=="TPolyLine") ||
-	       (myClass=="TText") 
+	       (myClass=="TText") ||
+	       (myClass=="TArc") || 
+	       (myClass=="TBox") 
 	       ) {
     } else {
-      logWARNING(Form("Unhandled class %s", myClass.c_str()));
+      std::cerr << Form("Unhandled class %s", myClass.c_str()) << std::endl;
     }
   }
 }
@@ -1080,9 +1085,9 @@ ostream& RootWTextFile::dump(ostream& output) {
   std::ofstream outputFile;
   string destinationFileName = targetDirectory_ +"/" + fileName_;
   outputFile.open(destinationFileName.c_str());
-  // TODO: add a check heer if the file was correctly opened
-  //outputFile << myText_.str();
-  //myText_ >> outputFile;
+  // TODO: add a check here if the file was correctly opened
+  // outputFile << myText_.str();
+  // myText_ >> outputFile;
   outputFile << myText_.str() << endl;
   outputFile.close();
 
@@ -1098,7 +1103,7 @@ ostream& RootWTextFile::dump(ostream& output) {
 //*******************************************//
 
 ostream& RootWBinaryFile::dump(ostream& output) {
-  if (originalFileName_=="") {
+  if ((originalFileName_=="")&&(!noCopy_)) {
     cerr << "Warning: RootWBinaryFile::dump() was called without prior setting the original file name" << endl;
     return output;
   } 
@@ -1109,7 +1114,7 @@ ostream& RootWBinaryFile::dump(ostream& output) {
 
   string destinationFileName = targetDirectory_ +"/" + fileName_;
 
-  if (boost::filesystem::exists(originalFileName_) && originalFileName_ != destinationFileName) { // CUIDADO: naive control on copy on itself. it only matches the strings, not taking into account relative paths and symlinks
+  if ((!noCopy_)&&(boost::filesystem::exists(originalFileName_) && originalFileName_ != destinationFileName)) { // CUIDADO: naive control on copy on itself. it only matches the strings, not taking into account relative paths and symlinks
     try {
       if (boost::filesystem::exists(destinationFileName))
         boost::filesystem::remove(destinationFileName);
@@ -1209,3 +1214,26 @@ ostream& RootWInfo::dump(ostream& output) {
          << value_ << "</a></tt><br/>";
   return output;
 }
+
+//*******************************************//
+// RootWGraphViz                             //
+//*******************************************//
+
+ostream& RootWGraphViz::dump(ostream& output) {
+  string myDescription = description_;
+  description_ += " (GraphViz)";
+  RootWTextFile::dump(output);
+  int result = system("which dot > /dev/null");
+  if (result==0) {
+    string svgFileName = fileName_+".svg";
+    string command = "dot -Tsvg " + targetDirectory_ +"/" + fileName_ + " > " + targetDirectory_ + "/" + svgFileName;
+    result = system(command.c_str());
+    if (result==0) {
+      RootWBinaryFile* mySvg = new RootWBinaryFile(svgFileName, myDescription+" (SVG)" );
+      mySvg->setNoCopy(true);
+      mySvg->dump(output);
+    }
+  }
+  return output;
+}
+
